Fixed update, added protect routes, styling

+++ flask_app/controllers/lists.py
@@ -7,11 +7,11 @@ from flask import flash
 @app.route("/shopping_list")
 def shopping_list():
     if "user_id" not in session:
-        flash("You must be signed in to access the dashboard.", "Signin")
+        flash("You must be signed in to access your shopping list.", "Signin")
         return redirect("/")
 
     user = User.get_user_id(session["user_id"])
-    list = List.get_all()
+    list = List.create_user_list(session["user_id"])
     return render_template("shopping_list.html", user=user, list=list)
 
 @app.route("/shopping_list/<int:list_id>")
@@ -27,6 +27,9 @@ def list_create_page():
 
 @app.route("/item/edit/<int:list_id>")
 def list_edit_page(list_id):
+    if "user_id" not in session:
+        flash("You must be signed in to edit items.", "Signin")
+        return redirect("/")
     user = User.get_user_id(session["user_id"])
     list = List.get_by_id(list_id)
     return render_template("edit.html", user=user, list=list)
@@ -35,18 +38,13 @@ def list_edit_page(list_id):
 def create_list():
         valid_list = List.create_valid_list(request.form)
         if valid_list:
-            return redirect(f'/shopping_list/{valid_list.id}')
-        return redirect(f'/item/create')
-
-@app.route("/shopping_list/<int:list_id>", methods=["POST"])
-def update_list(list_id):
-    valid_list = List.update_list(request.form, session["user_id"])
-
-    if not valid_list:
-        return redirect(f"/item/edit/{list_id}")
-    return redirect(f"/item/edit/{list_id}")
+            return redirect(f'/shopping_list')
+        return redirect(f'/item/add')
 
 @app.route("/item/delete/<int:list_id>")
 def delete_by_id(list_id):
+    if "user_id" not in session:
+        flash("You must be signed in to delete itmes from your list.", "Signin")
+        return redirect("/")
     List.delete_list_by_id(list_id)
     return redirect("/shopping_list")
